+++ src/main/java/chess/controller/ChessController.java
@@ -5,8 +5,7 @@ import chess.board.File;
 import chess.board.Position;
 import chess.board.Rank;
 import chess.board.dto.BoardDto;
-import chess.game.GameCommand;
-import chess.game.GameStatus;
+import chess.view.GameCommand;
 import chess.piece.AllPiecesGenerator;
 import chess.piece.Pieces;
 import chess.piece.Side;
