+++ src/main/java/chess/view/InputView.java
@@ -1,6 +1,5 @@
 package chess.view;
 
-import chess.game.GameCommand;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
